Update file
I fix some translations and add the last entries. There are some entries that may need further adjustments because in Spanish there are different uses. Vg fixed is 'fijo' per se, but something like 'fixed hours' is 'horas fijas'. In Spanish words have gender and number and the nouns and the articles and adjectives must be of the same gender and number. Besides that issue, some words have several meanings, using the same example word 'fixed', that word means by opposite of flexible something, but as 'repaired' something else...